logrotate: fixes for systemd
This is a combination of 4 commits:

3109a8855291b7ee4270c72cf92d1a44b9ae2692 logrotate: fixes for systemd

Before this patch, the command 'logrotate -f /etc/logrotate.d/ceph'
was generating an error "Failed to reload ceph.target: Job type reload is not
applicable for unit ceph.target".

https://bugzilla.suse.com/show_bug.cgi?id=931451 Fixes: bnc#931451

Signed-off-by: Nathan Cutler <ncutler@suse.com>

d35cbfe411704d41c2cdada7e24fe9c09acbe956 logrotate: check for active systemd services

Before we issue systemctl reload, check that there is at least
one active ceph* service.

Signed-off-by: Nathan Cutler <ncutler@suse.com>

1e85f63185b2a39dc1f613468f4bd5d95f6e9654 src/logrotate.conf: add hyphen to systemctl reload

Without the hyphen, the systemctl reload command matches ceph.target,
which generates an undesirable error message.

Signed-off-by: Nathan Cutler <ncutler@suse.com>

145364bbd933da8c1f384bea41f1b5cf37e9e082 logrotate: fix log rotation with systemd

systemctl does not have a nice way to enumerate (active) units so we can
reload them.  On centos7, the is-active wildcard syntax does not appear to
be supported.  On fedora 22, it prints the state only but not which unit
the state belongs to.

Simply killall -1 instead.

Signed-off-by: Sage Weil <sage@redhat.com>

Conflicts:
	src/logrotate.conf
	src/rgw/logrotate.conf
            Ignore some Debian-specific differences