ceph.spec.in:BuildRequires sharutils
sharutils rpm as a BuildRequires on fedora breaks the open build service which
is used by suse. rpm conditionals are confusing. What was previously here was
very confusing. This build dependency was set due to the confusing nature of
spec files so that the build dependency was on all platforms that are not rhel.

sharutils is needed if you build with java enabled

It is included in:
    m4/ac_check_class.m4
    m4/ac_prog_java_works.m4

Both these are only executed if cephfs_java is enabled.

Thanks to Nathan Cutler <ncutler@suse.cz> for going into the
details with me.

Signed-off-by: Owen Synge <osynge@suse.com>