ceph.spec.in: specify local sources
The Factory checker attempts to download the tarball to compare it
bit-for-bit with what we have in the package, but the expanded URL is not
found on the remote server.

The Factory checker also demands that ceph-rpmlintrc be listed as a
source file.

Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit 741083c3bf281299fe43fc08ca3acdc8ada7b190)