systemd: reinstate templated ceph-osd@.service.in
On RH and derivatives, %{_libexecdir} expands to /usr/libexec, while on
SUSE derivatives it expands to /usr/lib.

This was addressed in ac347dc3402a50fa0ba8c04cf2f93c36283eca87
which templated the ceph-osd@.service using autotools.

The crucial line containing the template variable was (inadvertently)
removed in 7c9fdf44f2c18659a0bcc03f7b98dafdf9f54448 and from there it was a
logical step to remove the template altogether in
367c794cb1ab701d6a28bdbaa5f7cd84e16a5e32

However, this breaks SUSE builds, requiring a downstream patch, of which
the template was designed to be a more robust replacement.

http://tracker.ceph.com/issues/14705 Fixes: #14705

Signed-off-by: Nathan Cutler <ncutler@suse.com>

Conflicts:
	systemd/ceph-osd@.service.in